+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - The table field of the content element was moved from FlexForms to a native database field
 
+### Fixed
+- Connection not shown in data overview when disabled (#27)
+
 ### Removed
 - Compatibility with TYPO3 v12 (#26)