+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.0] - 2026-04-22
+
 ### Added
 - Compatibility with TYPO3 v14
 
@@ -277,7 +279,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release
 
-[Unreleased]: https://github.com/jobrouter/typo3-data/compare/v4.0.0...HEAD
+[Unreleased]: https://github.com/jobrouter/typo3-data/compare/v5.0.0...HEAD
+[5.0.0]: https://github.com/jobrouter/typo3-data/compare/v4.0.0...v5.0.0
 [4.0.0]: https://github.com/jobrouter/typo3-data/compare/v3.0.2...v4.0.0
 [3.0.2]: https://github.com/jobrouter/typo3-data/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/jobrouter/typo3-data/compare/v3.0.0...v3.0.1